% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isys.R
\name{effort.is}
\alias{effort.is}
\title{Implementation system module to set fishing mortalities relative to a (dynamic)  reference value}
\usage{
effort.is(
  stk,
  ctrl,
  Fref = yearMeans(fbar(stk)[, ac(seq(dy - nyears, dy))]),
  nyears = args$nsqy,
  args,
  tracking
)
}
\arguments{
\item{stk}{The observed or estimated FLStock.}

\item{ctrl}{A fwdControl object setting an F or effort target.}

\item{Fref}{Reference value over which the multiplier is computed. Defaults to an average over the last observed nyears}

\item{nyears}{Number of years to use to compute reference value. strating on last data year (ay - data_lag).}

\item{args}{A list containing dimensionality arguments, passed on by mp().}

\item{tracking}{An FLQuant used for tracking indicators, intermediate values, and decisions during MP evaluation.}
}
\description{
Implementation system module to set fishing mortalities relative to a (dynamic)  reference value
}
\examples{
# Example dataset
data(plesim)

# Sets up an mpCtrl using hockeystick(fbar~ssb)
ctrl <- mpCtrl(
  est = mseCtrl(method=perfect.sa),
  hcr = mseCtrl(method=hockeystick.hcr, args=list(metric="ssb", trigger=45000, 
    output="fbar", target=0.27)),
  isys = mseCtrl(method=effort.is, args=list(nyears=3)))

# Runs mp between 2021 and 2035
run <- mp(om, control=ctrl, args=list(iy=2021, fy=2035))

# Runs mp without effort.is 'nyears' buffer effect
run_nois <- mp(om, control=ctrl[-3], args=list(iy=2021, fy=2035))

# Plots results
plot(om, effort.is=run, no_is=run2)
}
