% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{inspect}
\alias{inspect}
\title{Inspect MSE Tracking Data}
\usage{
inspect(tab, metrics = NULL, summary = medmad)
}
\arguments{
\item{tab}{An \code{FLms+e} object or a \code{data.table} containing tracking data. If an \code{FLmse} object is supplied, the \code{tracking} slot is extracted.}

\item{metrics}{A character vector specifying the metrics to include in the output. If a single character value is given and it does not match excatly any of the contained metrics, it is used to subset using the datatable::\%ilike\% function. The special value \code{"decisions"} can be used to select all metrics from \code{"hcr"} onward. If \code{NULL}, the default, all metrics are returned.}

\item{summary}{A function (such as \code{mean} or \code{median}) to summarize the \code{data} column within the tracking \code{data.table} across rthe \code{iter} dimension. Defaults to \code{medmad} which returns a string with \code{"Median (Median Absolute Deviation)"}.}
}
\value{
A \code{data.table} where rows correspond to years and columns to selected metrics, with the tracking \code{data} aggregated using the specified summary function.
}
\description{
The \code{inspect} function turns the long \code{tracking} table into a year a metric summary
on which to track the inoputs and outputs of all steps inside a call to the \code{mp}
function.
}
\details{
The function processes the table on the \code{tracking} slot by:
\itemize{
\item Arranging metrics based on the order in which they are produced inside \code{mp()}.
\item Subsets metrics based on user input or predefined criteria like \code{"decisions"}.
\item Aggregates the \code{data} column using the \code{summary} function provided.
\item Reshapes the processed \code{data.table} into a year-by-metrics format for easy reading.
}

The predefined metric order is that of the various steps inside \code{mp()}: \code{"om"}, \code{"obs"}, \code{"est"}, \code{"ind"}, \code{"phcr"}, \code{"hcr"}, \code{"isys"}, \code{"tm"}, \code{"iem"}, \code{"fb"}, \code{"fwd"}. Extra tracks produced inside any module are placed before the one named after
the module that produced it.
}
\examples{
# Load example MP run
data(mserun)
# Inspect all metrics
inspect(run)
# Inspect with different summary function
inspect(tracking(run), summary=mean)
# Inspect metrics with "SB" or "sb" in their names
inspect(tracking(run), metrics="SB")
# Inspect specific metrics
inspect(tracking(run), metrics=c("C.om", "C.obs", "C.est"))
# Inspect only decision-related metrics from 'hcr'
inspect(tracking(run), metrics="decisions")
}
