% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{setLabelPerformance}
\alias{setLabelPerformance}
\title{setLabelPerofrmance}
\usage{
setLabelPerformance(file = "model/performance.dat.gz", labels)
}
\arguments{
\item{file}{File where performance table is stored, default: 'model/performance.dat.gz'}
}
\value{
Invisible updates the table in file
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
