% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{labelPerformance}
\alias{labelPerformance}
\title{labelPerformance}
\usage{
labelPerformance(dat, labels = NULL)
}
\arguments{
\item{dat}{A performance statistics table, as returned by performance()}

\item{labels}{Labels to be inserted, as vector or data.frame/data.table}
}
\value{
A labelled performance statistics data.table
}
\description{
Creates a label column in a performance table
}
\details{
\itemize{
\item 'numeric'
\item vector
\item data.frame or data.table
\item NULL
}
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
