% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performanceFLQuant}
\alias{performanceFLQuant}
\title{Convert Performance Data to FLQuant Format}
\usage{
performanceFLQuant(x)
}
\arguments{
\item{x}{An object of class \code{FLmse}, \code{FLmses}, or a \code{data.table}.}
}
\value{
If a single management procedure (mp) is present, returns a single \code{FLQuant}
object. If multiple management procedures are present, returns a list of
\code{FLQuant} objects, one per mp. When \code{biol} column is present, it is renamed
to \code{unit} in the resulting \code{FLQuant} or \code{FLQuants}.
}
\description{
Transforms performance metrics from \code{FLmse} or \code{FLmsea} objects, stored as a
data.table into an \code{FLQuant} pr \code{FLQuants} object.
}
\details{
If \code{x} is of class \code{FLmse} or \code{FLmses}, the performance slot is extracted
automatically. The data.table must contain columns: \code{statistic}, \code{year}, \code{iter},
\code{data}, and \code{mp}. Optionally, a \code{biol} column can be present, as in the
performance table obtained from callin \code{mp()} on an \code{FLombf} OM. Values for each
\code{biol} will be separated using the 'unit' dimension.

The first (\code{quant}) dimension of the \code{FLQuant} or \code{FLQuants} contain the
statistics, named as in the \code{statistic} column of the performance table.

When results from multiple management procedures are present, the output is an
\code{FLQuants} list of \code{FLQuant} objects, one per mp.
}
\examples{
\dontrun{
  # Using a data.table directly
  perf_dt <- data.table(
    statistic = rep("SSB", 4),
    year = rep(2025:2026, 2),
    iter = rep(1:2, each = 2),
    data = runif(4, 1000, 3000),
    mp = "MP1"
  )
  quants <- performanceFLQuants(perf_dt)
}

}
\seealso{
\code{\link[=performance]{performance()}}, \link[mse:FLmse-class]{mse::FLmse}, \linkS4class{FLmses}
}
\keyword{manip}
