% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{pid.hcr}
\alias{pid.hcr}
\title{Proportional-Integral-Derivative (PID) Harvest Control Rule (HCR)}
\usage{
pid.hcr(
  stk,
  ind,
  ref,
  metric = ssb,
  initial,
  kp = 0,
  ki = 0,
  kd = 0,
  nyears = 5,
  dlow = NA,
  dupp = NA,
  args,
  tracking,
  ...
)
}
\arguments{
\item{stk}{\code{FLStock}. The stock object to which the HCR applies.}

\item{ind}{\code{FLQuant}. The abundance index used to compute the control signal.}

\item{ref}{Numeric. The reference value for the metric to determine divergence.}

\item{metric}{Character or function. The metric applied to measure stock status. Default is \code{'ssb'} (spawning stock biomass).}
}
\description{
Implements a Proportional-Integral-Derivative (PID) based harvest control rule for adjusting
Total Allowable Catch (TAC) based on divergence from a reference point using PID control signals.
}
