% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isys.R
\name{bank_borrow.is}
\alias{bank_borrow.is}
\title{bank_borrow.isys}
\usage{
bank_borrow.is(
  stk,
  ctrl,
  args,
  split = NULL,
  rate = NULL,
  diff = 0.15,
  healthy = 1,
  tracking
)
}
\arguments{
\item{stk}{\code{FLStock} object representing the stock dat.}

\item{ctrl}{\code{fwdControl} object for forward projection controls.}

\item{args}{List of additional arguments required for the execution.}

\item{split}{Optional. Input for splitting controls (default is \code{NULL}).}

\item{rate}{Numeric. Proportion rate for borrowing and banking adjustments (default is \code{NULL}).}

\item{diff}{Numeric. Tolerance difference for determining borrowing or banking actions (default is 0.15).}

\item{healthy}{Numeric. Threshold for determining stock health (default is 3).}

\item{tracking}{Data structure. Object to track borrowing and banking values over time.}
}
\value{
A list containing:
\item{ctrl}{Modified \code{fwdControl} object with updated TAC values.}
\item{tracking}{Updated tracking data structure with recorded borrowing and banking amounts.}
}
\description{
Banking and Borrowing Mechanism for TAC Adjustment
}
\details{
The function performs the following:
\itemize{
\item Checks if management assumes annual (frq = 1). Raises an error if not.
\item Retrieves and adjusts TAC for previous banking and borrowing actions.
\item Verifies the stock's health status (\code{healthy}) using tracking data to determine eligibility for banking/borrowing.
\item Calculates borrowing if TAC is significantly lower than the reference (\code{pre}) and banking if TAC is significantly higher.
\item Updates the \code{ctrl} and \code{tracking} objects with the adjusted TAC and banking/borrowing amounts.
}
}
\note{
Banking and borrowing adjustments are applied only when the frequency (\code{frq}) is equal to 1.
}
\examples{
# Example dataset
data(sol274)

# Sets up an mpCtrl using hockeystick(catch~ssb) + bank_borrow(10\%)
ctrl <- mpCtrl(est = mseCtrl(method=perfect.sa),
  hcr = mseCtrl(method=hockeystick.hcr, args=list(metric="ssb", trigger=42000, 
    output="catch", target=11000)),
  isys = mseCtrl(method=bank_borrow.is, args=list(rate=0.10, healthy=2, diff=0.05)))

# Runs the MP
run <- mp(om, control=ctrl, args=list(iy=2021, fy=2035))

# Plot time series
plot(om, list(BaB=run))

# Observe from tracking the TAC-setting steps
items <- c("year", "hcr", "banking.isys", "borrowing.isys", "isys", "fwd")

dcast(tracking(run)[metric \%in\% items[-1], .(data=mean(data)), by=.(year, metric)],
  year~metric, value.var='data')[, ..items]
}
