% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{buffer.hcr}
\alias{buffer.hcr}
\alias{plot_buffer.hcr}
\title{Buffer Harvest Control Rule (HCR)}
\usage{
buffer.hcr(
  stk,
  ind,
  metric = "wmean",
  target = 1,
  width = 0.5,
  lim = max(target * 0.1, target - 2 * width),
  bufflow = max(lim * 1.5, target - width),
  buffupp = target + width,
  sloperatio = 0.15,
  initial,
  nyears = 4,
  dupp = NULL,
  dlow = NULL,
  all = TRUE,
  scale = FALSE,
  ...,
  args,
  tracking
)

plot_buffer.hcr(
  args,
  obs = "missing",
  alpha = 0.3,
  labels = c(lim = "limit", bufflow = "Lower~buffer", buffupp = "Upper~buffer", metric =
    metric, output = output),
  metric = args$metric,
  output = "multiplier",
  xlim = buffupp * 1.5,
  ylim = scale * 1.5
)
}
\arguments{
\item{stk}{The 'oem' observation or SA estimation, an FLStock object.}

\item{ind}{Possible indicators returned by the 'est' step, FLQuants.}

\item{metric}{The metric name used for the x-axis label.}

\item{target}{The desired target value for the metric, defaults to 1. Numeric or FLQuant.}

\item{width}{Numeric. The width of the buffer zone surrounding the target. Default is \code{0.5}.}

\item{lim}{Numeric. The lower threshold beyond which the HCR steeply reduces catch. Default is \code{max(target * 0.10, target - 2 * width)}.}

\item{bufflow}{Numeric. The lower bound of the buffer range. Default is \code{max(lim * 1.50, target - width)}.}

\item{buffupp}{Numeric. The upper bound of the buffer range. Default is \code{target + width}.}

\item{sloperatio}{Numeric. Defines the slope for values exceeding the buffer's upper limit. Default is \code{0.15}.}

\item{initial}{Numeric or \code{FLQuant}. The previous year's TAC value or another scaling factor for outputs.}

\item{nyears}{Numeric. Number of years to use for windowing or smoothing metrics. Default is 4.}

\item{dupp}{A limit for the increase in the output variable, e.g. 1.15 for a maximum increase of 15\%, numeric. No limit is applied if NULL.}

\item{dlow}{A limit for the decrease in the output variable, e.g. 0.85 for a maximum decrease of 15\%, numeric. No limit is applied if NULL.}

\item{all}{If \code{TRUE}, upper and lower limits (\code{dupp} and \code{dlow}) are applied unconditionally, otherwise only when metric > bufflow, logical.}

\item{...}{Any extra arguments to be passed to the function computing 'metric'.}

\item{args}{list. The list of parameters used by the buffer.hcr function.}

\item{tracking}{A data.table used for tracking indicators, intermediate values, and decisions during MP evaluation.}

\item{obs}{Observed values to overlay on the plot.}

\item{alpha}{Alpha transparency for observed points/lines.}

\item{labels}{Labels for the plot annotations. A list with names among 'lim', 'bufflow', 'buffupp', 'metric', and 'output'.}

\item{output}{The output name used for the y-axis label.}

\item{xlim}{The maximum x-axis limit for the plot.}

\item{ylim}{The maximum y-axis limit for the plot.}
}
\value{
ggplot2 object
}
\description{
Implements a "buffer" style harvest control rule (HCR) to set changes in catch levels by using a given metric, such as the value of an index of abundance, or an estimate of biomass. This is compared with some defined buffer and limit reference points, to compute a multiplier that will set future catches from previous levels.
with optional bounds on TAC changes.
}
\details{
\code{plot_buffer.hcr} Plots the buffer harvest control rule (HCR) curve along with
optional observed values.
}
