% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{depletion.hcr}
\alias{depletion.hcr}
\title{Depletion Harvest Control Rule (HCR)}
\usage{
depletion.hcr(
  stk,
  ind,
  metric = "ssb",
  mult = 1,
  hrmsy,
  K,
  trigger = 0.4,
  lim = 0.1,
  min = 1e-05,
  dupp = NULL,
  dlow = NULL,
  all = TRUE,
  ...,
  args,
  tracking
)
}
\arguments{
\item{stk}{\code{FLStock}. The stock object to which the HCR applies.}

\item{ind}{\code{FLQuant}. The abundance index used as input for computing adjustments.}

\item{metric}{Character or function. The metric applied to measure stock status. Default is \code{'ssb'} (spawning stock biomass).}

\item{mult}{Numeric. A scaling multiplier for adjusting the harvest rate. Default is \code{1}.}

\item{hrmsy}{Numeric. Harvest rate that achieves maximum sustainable yield (MSY}
}
\description{
Implements a depletion-based harvest control rule (HCR), adjusting harvest rates and TAC (Total Allowable Catch)
based on stock status relative to depletion thresholds.
}
