% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isys.R
\name{effort.is}
\alias{effort.is}
\title{Calculate Effort Multiplier for Harvest Control Rules}
\usage{
effort.is(
  stk,
  ctrl,
  nyears = args$nsqy,
  Fref = yearMeans(fbar(stk)[, ac(seq(dy - nyears + 1, dy))]),
  args,
  tracking
)
}
\arguments{
\item{stk}{An FLStock object containing stock information}

\item{ctrl}{A fwdControl object setting an F or effort target.}

\item{nyears}{Integer. Number of years to use for calculating mean Fref.
Default is taken from \code{args$nsqy}}

\item{Fref}{Numeric or FLQuant. Reference fishing mortality for scaling. Defaults to
the mean of fbar over the specified years.}

\item{args}{A list containing dimensionality arguments, passed on by mp().}

\item{tracking}{An FLQuant used for tracking indicators, intermediate values, and decisions during MP evaluation.}
}
\value{
A list containing:
\describe{
\item{ctrl}{Updated control list with effort multiplier in \code{value}
and adjusted years in \code{rel.year}}
\item{tracking}{Updated tracking object}
}
}
\description{
Converts a harvest control rule (HCR) fihsing mortality target into a multiplier
relative to a (dynamic) reference fishing mortality level.
}
\details{
The function calculates the effort multiplier as:
\deqn{mult = \frac{target}{Fref}}{mult = target / Fref}
}
\note{
The \code{ctrl$quant} parameter must be one of "f", "fbar", or "effort".
Other values will cause the function to stop with an error.
}
\examples{
# Example dataset
data(plesim)

# Sets up an mpCtrl using hockeystick(fbar~ssb)
ctrl <- mpCtrl(
  est = mseCtrl(method=perfect.sa),
  hcr = mseCtrl(method=hockeystick.hcr, args=list(metric="ssb", trigger=45000, 
    output="fbar", target=0.27)),
  isys = mseCtrl(method=effort.is, args=list(nyears=3)))

# Runs mp between 2021 and 2035
run <- mp(om, control=ctrl, args=list(iy=2021, fy=2035))

# Runs mp without effort.is 'nyears' buffer effect
run_nois <- mp(om, control=ctrl[-3], args=list(iy=2021, fy=2035))

# Plots results
plot(om, effort.is=run, no_is=run_nois)
}
