% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{inspect}
\alias{inspect}
\title{Inspect MSE Tracking Data}
\usage{
inspect(tab, metrics = NULL, summary = medmad)
}
\arguments{
\item{tab}{An \code{FLmse} object or a \code{data.table} containing tracking data. If an \code{FLmse} object is supplied, the \code{tracking} slot is extracted.}

\item{metrics}{A character vector specifying the metrics to include in the output. If a single character value is given and it does not match excatly any of the contained metrics, it is used to subset using the datatable::\%ilike\% function. The special value \code{"decisions"} can be used to select all metrics from \code{"hcr"} onward. If \code{NULL}, the default, all metrics are returned.}

\item{summary}{A function (such as \code{mean} or \code{median}) to summarize the \code{data} column within the tracking \code{data.table} across rthe \code{iter} dimension. Defaults to \code{medmad} which returns a string with  \code{"Median (Median Absolute Deviation)"}.}
}
\value{
A \code{data.table} where rows correspond to years and columns to selected metrics, with the tracking \code{data} aggregated using the specified summary function.
}
\description{
The \code{inspect} function is designed to extract and format the tracking data from an \code{FLmse} object, performing aggregations and subsetting based on specified metrics and a summary function.
}
\details{

}
\examples{
# Load example MP run
data(mserun)
# Inspect all metrics
inspect(run)
# Inspect with different summary function
inspect(tracking(run), summary=mean)
# Inspect metrics with "SB" or "sb" in their names
inspect(tracking(run), metrics="SB")
# Inspect specific metrics
inspect(tracking(run), metrics=c("C.om", "C.obs", "C.est"))
# Inspect only decision-related metrics from 'hcr'
inspect(tracking(run), metrics="decisions")
}
